refactor(graph): split budget and process test modules into sibling files
Server Guards enforces MAX_LINES=1500 AND MAX_BYTES=51200; the new tests pushed
`budget.rs` to 73293 bytes and `process.rs` to 57070 bytes — the latter under the
line limit but over the byte limit.

Both test modules move verbatim into sibling files included with `#[path]`, the
same convention `process_linux_supervisor_tests.rs` and `graph_warmer_ledger_tests.rs`
already use, so they stay children of the module under test and keep their
`use super::*` access to private internals. No production logic is touched and no
`djinn:allow-oversize` marker is added — that escape hatch is for pre-existing
bulk, not for files a change just grew.

  budget.rs   73293 -> 32582 bytes   (+ budget_tests.rs   37504)
  process.rs  57070 -> 45847 bytes   (+ process_tests.rs  10733)

`scripts/check-file-size.sh` now reports 0 oversized files. Re-verified by
neutralization after the split: reverting `adaptive_ceiling` to the flat
`max_cap * 3` still fails 9 tests, so the split did not sever the linkage.
